changes before nap
NOTE ON THE SELECTION BUG:
pretty sure it's a rounding error. you must click in the upper right 25%
of a piece to select it.  I know brad did some offset thing to
compensate for this, but i wasnt quite sure how it worked so I couldn't
properly apply it.  If anyone wants to fix that, I don't think it's too
complicated.  I would start by looking at lines 76 and 77 of board.cs