Some changes so Kirsten can start working.
So what I want is for it to be possible for a selected piece to move.
On line 134 there is an if statement "if(selectedPiece == "none")" which
is the part im working on right now.  I would like to have something
that should start with "if(selectedPiece == "whiteKing") that should
handle the ability of the player to make the king move after selecting
them